Refactor exception analysis into new class
