+++ src/main/user-impl/java/com/mysql/cj/jdbc/ha/plugins/AutoCommitOffTransactionBoundaryState.java
@@ -33,23 +33,24 @@ import com.mysql.cj.jdbc.JdbcConnection;
 
 public class AutoCommitOffTransactionBoundaryState implements IState {
 
-    private final ConnectionMethodAnalyzer analyzer = new ConnectionMethodAnalyzer();
+    private final ConnectionMethodAnalyzer methodAnalyzer = new ConnectionMethodAnalyzer();
+    private final ExceptionAnalyzer exceptionAnalyzer = new ExceptionAnalyzer();
 
     @Override
     public IState getNextState(JdbcConnection currentConnection, String methodName, Object[] args) {
-        if (analyzer.isSetAutoCommitTrue(methodName, args)) {
+        if (methodAnalyzer.isSetAutoCommitTrue(methodName, args)) {
             return ReadWriteSplittingStateMachine.AUTOCOMMIT_ON_STATE;
         }
 
-        if (analyzer.isMethodClosingTransaction(methodName, args)) {
+        if (methodAnalyzer.isMethodClosingTransaction(methodName, args)) {
             return ReadWriteSplittingStateMachine.AUTOCOMMIT_OFF_TRANSACTION_BOUNDARY_STATE;
         }
 
-        if (analyzer.isSetReadOnlyFalse(methodName, args)) {
+        if (methodAnalyzer.isSetReadOnlyFalse(methodName, args)) {
             return ReadWriteSplittingStateMachine.READ_WRITE_STATE;
         }
 
-        if (analyzer.isExecuteDml(methodName, args) || analyzer.isExecuteStartingTransaction(methodName, args)) {
+        if (methodAnalyzer.isExecuteDml(methodName, args) || methodAnalyzer.isExecuteStartingTransaction(methodName, args)) {
             return ReadWriteSplittingStateMachine.AUTOCOMMIT_OFF_TRANSACTION_STATE;
         }
 
@@ -58,11 +59,11 @@ public class AutoCommitOffTransactionBoundaryState implements IState {
 
     @Override
     public IState getNextState(Exception e) {
-        if (analyzer.isFailoverException(e)) {
+        if (exceptionAnalyzer.isFailoverException(e)) {
             return ReadWriteSplittingStateMachine.AUTOCOMMIT_OFF_STATE;
         }
 
-        if (analyzer.isCommunicationsException(e)) {
+        if (exceptionAnalyzer.isCommunicationsException(e)) {
             return ReadWriteSplittingStateMachine.AUTOCOMMIT_OFF_TRANSACTION_BOUNDARY_STATE;
         }
